requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg[binary]==3.2.1
python-dotenv==1.0.1
apscheduler==3.10.4
pytest
pytest-cov
Faker
factory_boy

